Moving ITypeNode interface into an interfaces file
