--- src/nodes/ITypeNode.ts
@@ -1,6 +1,0 @@
-import * as ts from 'typescript'
-
-export default interface ITypeNode {
-  toEnum(): string
-  toAST(): ts.TypeNode
-}
