+++ src/nodes/interfaces.ts
@@ -1,0 +1,6 @@
+import { TypeNode } from 'typescript'
+
+export interface ITypeNode {
+  toEnum(): string
+  toAST(): TypeNode
+}
