+++ src/read-types.ts
@@ -3,7 +3,7 @@ import * as ts from 'typescript'
 import { read as eRead } from './ast/enum-mapped'
 import { identifiers as _id } from './ast/identifiers'
 import { methods as _methods } from './ast/methods'
-import ITypeNode from './nodes/ITypeNode'
+import { ITypeNode } from './nodes/interfaces'
 
 // Map/Set/List don't seem to use the etype,ktype,vtype property that's initialized
